config: lx2162aqds: use USXGMII ports as TYPE_FIXED
SERDES protocols which have XFI/USXGMII protocols on some lanes are
tested using the M11 mezzanine cards which do not have PHYs on the, thus
Linux configuration is not needed necessary. Just put the DPMACs as
TYPE_FIXED so that we do not need to change the kernel DTS for each
SERDES protocols adding a fixed-link node and changing the protocol
used.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>