config: add the second 100G port (MAC.2) for the LX2160AQDS RCW protocol 13
The second 100G port was disabled from the DPC file for the LX2160AQDS
RCW protocol 13. Enable it back so that it's available for usage.

Please note that no DPNI is automatically created and connected to the
DPMAC.2 from the DPL, the user needs to create it dynamically once the
board is booted.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>